move qrels to new module eval
